# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vivid OPS API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiErrorBody {
  error: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  "super-admin": "super-admin",
  admin: "admin",
  supervisor: "supervisor",
  user: "user",
} as const;

export type UserStatus = (typeof UserStatus)[keyof typeof UserStatus];

export const UserStatus = {
  active: "active",
  inactive: "inactive",
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  status: UserStatus;
  mustResetPassword: boolean;
  /** @nullable */
  lastSignInAt?: string | null;
  createdAt: string;
}

export type UserInputDelivery =
  (typeof UserInputDelivery)[keyof typeof UserInputDelivery];

export const UserInputDelivery = {
  "email-invite": "email-invite",
  "temp-password": "temp-password",
} as const;

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: UserRole;
  delivery: UserInputDelivery;
}

export interface UserUpdate {
  /** @minLength 1 */
  name?: string;
  email?: string;
  role?: UserRole;
  status?: UserStatus;
}

export type CreateUserResultDelivery =
  (typeof CreateUserResultDelivery)[keyof typeof CreateUserResultDelivery];

export const CreateUserResultDelivery = {
  "email-invite": "email-invite",
  "temp-password": "temp-password",
} as const;

export interface CreateUserResult {
  user: User;
  delivery: CreateUserResultDelivery;
  /**
   * Returned only when delivery is "temp-password"
   * @nullable
   */
  tempPassword?: string | null;
  /**
   * True when an invite email was successfully dispatched
   * @nullable
   */
  emailSent?: boolean | null;
}

export interface LoginInput {
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface LoginResult {
  user: User;
}

export interface ResetPasswordInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}