# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Vivid OPS API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
  - name: auth
  - name: users
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Sign in with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResult"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /auth/logout:
    post:
      operationId: logout
      tags: [auth]
      summary: Sign out the current session
      responses:
        "204":
          description: Logged out

  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Current authenticated user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not signed in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiErrorBody"

  /auth/reset-password:
    post:
      operationId: resetPassword
      tags: [auth]
      summary: Set a new password (used after first sign-in with temp password)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordInput"
      responses:
        "200":
          description: Password reset
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List users (admin / super-admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [users]
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResult"

  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get a single user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update a user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete a user
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "204":
          description: Deleted

  /users/{id}/resend-invite:
    post:
      operationId: resendInvite
      tags: [users]
      summary: Resend the invite email (and generate a new temp password)
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Invite resent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CreateUserResult"

components:
  schemas:
    HealthStatus:
      type: object
      required: [status]
      properties:
        status: { type: string }

    ApiErrorBody:
      type: object
      required: [error]
      properties:
        error: { type: string }

    UserRole:
      type: string
      enum: [super-admin, admin, supervisor, user]

    UserStatus:
      type: string
      enum: [active, inactive]

    User:
      type: object
      required: [id, email, name, role, status, mustResetPassword, createdAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        name: { type: string }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }
        mustResetPassword: { type: boolean }
        lastSignInAt:
          type: ["string", "null"]
          format: date-time
        createdAt: { type: string, format: date-time }

    UserInput:
      type: object
      required: [name, email, role, delivery]
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/UserRole" }
        delivery:
          type: string
          enum: [email-invite, temp-password]

    UserUpdate:
      type: object
      properties:
        name: { type: string, minLength: 1 }
        email: { type: string, format: email }
        role: { $ref: "#/components/schemas/UserRole" }
        status: { $ref: "#/components/schemas/UserStatus" }

    CreateUserResult:
      type: object
      required: [user, delivery]
      properties:
        user: { $ref: "#/components/schemas/User" }
        delivery:
          type: string
          enum: [email-invite, temp-password]
        tempPassword:
          type: ["string", "null"]
          description: Returned only when delivery is "temp-password"
        emailSent:
          type: ["boolean", "null"]
          description: True when an invite email was successfully dispatched

    LoginInput:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 1 }

    LoginResult:
      type: object
      required: [user]
      properties:
        user: { $ref: "#/components/schemas/User" }

    ResetPasswordInput:
      type: object
      required: [currentPassword, newPassword]
      properties:
        currentPassword: { type: string, minLength: 1 }
        newPassword: { type: string, minLength: 8 }