# lib/api-zod/src/generated/types/resetPasswordInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vivid OPS API
 * OpenAPI spec version: 0.1.0
 */

export interface ResetPasswordInput {
  /** @minLength 1 */
  currentPassword: string;
  /** @minLength 8 */
  newPassword: string;
}