# lib/api-zod/src/generated/types/userInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Vivid OPS API
 * OpenAPI spec version: 0.1.0
 */
import type { UserInputDelivery } from "./userInputDelivery";
import type { UserRole } from "./userRole";

export interface UserInput {
  /** @minLength 1 */
  name: string;
  email: string;
  role: UserRole;
  delivery: UserInputDelivery;
}